+++ NEWS.md
@@ -1,7 +1,25 @@
-# AHPtools 0.3.0
+# AHPtools 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.
 
 - Corrected a bug in the revisedConsistency() function
 
 - Added triadic Preference Reversal based consistency functionality
+
+## New features
+
+Functions added to compute weights for alternatives and lowest level sub criteria in AHP responses 
+and to view the full AHP tree as a parent-child hierarchy
+
+### Features
+
+- Added two new functions 
+  - viewAHPtree
+    see vignette to view the full AHP tree based on user-specified parent child relations
+  - AHPweights
+    see vignette to to print the final weights for the lowest level sub-criteria and the alternatives
+
+### Improved documentation
+
+- Added a vignette to view all functions in the AHPtools package in a neat form
+  see vignette